Some Linux Polishing
* Some profiling scopes to debug editor startup times (mostly limited by DLL loading and Qt init code like icon loading).
* Fix various compiler warnings. Most notably moved EZ_DEBUG_BREAK call to `int3` into a function as it was encountered in constexpr context which spammed some C++20 warning in GCC.
* Re-enabled OZZ SIMD use on Linux.
* I can't be bothered to maintain 3 pretty printer implementations for ezEngine so I added a github skill so an LLM can implement and self-check gdb pretty printer implementations.

+++ Code/EditorPlugins/VisualScript/EditorPluginVisualScript/VisualScriptGraph/VisualScriptVariable.cpp
@@ -33,8 +33,8 @@ EZ_BEGIN_STATIC_REFLECTED_ENUM(ezVisualScriptVariableType, 1)
 EZ_END_STATIC_REFLECTED_ENUM;
 // clang-format on
 
-static_assert(ezVisualScriptVariableType::Variant == ezVisualScriptDataType::Variant);
-static_assert(ezVisualScriptVariableType::Resource == ezVisualScriptDataType::Resource);
+static_assert(static_cast<int>(ezVisualScriptVariableType::Variant) == static_cast<int>(ezVisualScriptDataType::Variant));
+static_assert(static_cast<int>(ezVisualScriptVariableType::Resource) == static_cast<int>(ezVisualScriptDataType::Resource));
 
 ///////////////////////////////////////////////////////////////////////////
